requirements.txt
celery==4.3.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Caching==1.8.0
Flask-Cors==3.0.8
Flask-HTTPBasicAuth==1.0.1
Flask-json-schema==0.0.5
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
jsonpickle==1.2
loguru==0.3.2
PyJWT==1.7.1
PyMySQL==0.9.3
pypinyin==0.36.0
redis==3.3.11
requests==2.22.0
simplejson==3.17.0
